+++ CHANGELOG.md
@@ -22,6 +22,7 @@ _Add new changes here_
 
 ### Added
 
+- Added Hugging Face model support to ClassifierContainer
 - Added `piffle` package as dependency ([#575](https://github.com/maps-as-data/MapReader/pull/575))
 
 ## [v1.8.1](https://github.com/Living-with-machines/MapReader/releases/tag/v1.8.1) (2025-08-11)